+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 For a list of breaking changes, check [here](#breaking-changes).
 
+## Unreleased
+- [#1653](https://github.com/clj-kondo/clj-kondo/issues/1653): new linter `:keyword-binding` warns when a keyword is used in a `:keys` binding vector. This linter is `:off` by default. See [docs](doc/linters.md#keyword-in-binding-vector).
+
 ## 2022.04.08
 
 - [#1331](https://github.com/clj-kondo/clj-kondo/issues/1331): new linter `:non-arg-vec-return-type-hint` that warns when a return type hint is not placed on the arg vector (CLJ only). See [docs](https://github.com/clj-kondo/clj-kondo/blob/master/doc/linters.md#non-arg-vec-return-type-hint).